make_know_pipe/requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM integrations
groq>=0.4.0
google-genai>=0.3.0

# Optional dependencies for enhanced functionality
PyPDF2>=3.0.0  # For PDF processing
Pillow>=10.0.0  # For image handling

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0